flask-app/requirements.txt
numpy==1.18.5
scikit-learn==0.23.1
pillow==7.2.0
Flask==1.1.2
requests==2.24.0